REPL simplification
The initialization piece of code does not have to have session id as part of it's name. It is just used to test wheter the compiler has been sucessfuly initialized. We don't therefore have to change the initialization code and thus can remove it.

Also the code in setupClassNames was simplified. This change also allows us to use the same code in this method in both spark with scala 2.10 and 2.11

Also remove unnecessary setting of classpath since it's always overriden in createInterpreter method.